Apps/hand-pose-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
python-multipart==0.0.6

# Phase 2: Add ONNX runtime for MediaPipe model
# onnxruntime-gpu==1.16.3  # For GPU support
# onnxruntime==1.16.3      # For CPU-only

